Extract edge, page_info and root field modules
