+++ lib/graphql/types/relay/base_edge.rb
@@ -21,40 +21,8 @@ module GraphQL
       #   end
       #
       # @see {Relay::BaseConnection} for connection types
-      class BaseEdge < Types::Relay::BaseObject
-        description "An edge in a connection."
-
-        class << self
-          # Get or set the Object type that this edge wraps.
-          #
-          # @param node_type [Class] A `Schema::Object` subclass
-          # @param null [Boolean]
-          def node_type(node_type = nil, null: true)
-            if node_type
-              @node_type = node_type
-              # Add a default `node` field
-              field :node, node_type, null: null, description: "The item at the end of the edge."
-            end
-            @node_type
-          end
-
-          def authorized?(obj, ctx)
-            true
-          end
-
-          def accessible?(ctx)
-            node_type.accessible?(ctx)
-          end
-
-          def visible?(ctx)
-            node_type.visible?(ctx)
-          end
-        end
-
-
-        field :cursor, String,
-          null: false,
-          description: "A cursor for use in pagination."
+      class BaseEdge < GraphQL::Schema::Object
+        include Types::Relay::EdgeBehaviors
       end
     end
   end
