+++ lib/graphql/types/relay/nodes_field.rb
@@ -21,25 +21,7 @@ module GraphQL
       #     end
       #   end
       #
-      NodesField = GraphQL::Schema::Field.new(
-        name: "nodes",
-        owner: nil,
-        type: [GraphQL::Types::Relay::Node, null: true],
-        null: false,
-        description: "Fetches a list of objects given a list of IDs.",
-        relay_nodes_field: true,
-      ) do
-        argument :ids, "[ID!]!", required: true,
-          description: "IDs of the objects."
-
-        def resolve(obj, args, ctx)
-          args[:ids].map { |id| ctx.schema.object_from_id(id, ctx) }
-        end
-
-        def resolve_field(obj, args, ctx)
-          resolve(obj, args, ctx)
-        end
-      end
+      NodesField = GraphQL::Schema::Field.new(**HasNodesField.field_options, &HasNodesField.field_block)
     end
   end
 end
